// src/main/resources/APachas/src/app/models/MUser.ts
export class MUser {
    userId: number;
    userName: string;
    userSurname: string;
    userLogin: string;
    userPassword: string;
    userEmail: string;
    userBirthday: Date;
    userPhoto: string;
    roles: string;
    permissions: string;
    userActive: boolean;
    userCreation: Date;
    userRemoval: Date;
    userNotify: boolean;
}